LayoutTests/fast/encoding/japanese-encoding-mix.html

<!doctype html public "-//W3C//DTD HTML 4.01 Transitional//EN">
<html dir="LTR" lang="ja">
<meta http-equiv="Content-Type" content="text/html; charset=EUC-JP">
<p>Some text here is encoded as EUC-JP, and some (in comment) as Shift_JIS. Since there is an explicit encoding declaration, auto-detection shouldn't change the encoding.</p><p>･ﾋ･ﾃ･ﾝ･･ﾆ･･ﾓ</p>
<!--右カラムコメントアウト ここから -->
<script>
if (window.testRunner)
    testRunner.dumpAsText();
document.write("Should be EUC-JP: " + (document.charset ? document.charset : document.characterSet));</script>
